requirements.txt
flask
uWSGI
requests
python-dotenv
openai >= 0.27.0
injector
redis
slack_bolt
